// packages/react/src/components/Box/types/box.d.ts
import type { CSSProperties, HTMLAttributes, ReactNode } from "react";

export type SpacingValueBox = number | string;

export type BoxProps = HTMLAttributes<HTMLElement> & {
  children?: ReactNode;
  width?: number | string;
  height?: number | string;
  minWidth?: number | string;
  maxWidth?: number | string;
  minHeight?: number | string;
  maxHeight?: number | string;
  display?: CSSProperties["display"];
  position?: CSSProperties["position"];
  overflow?: CSSProperties["overflow"];
  overflowX?: CSSProperties["overflowX"];
  overflowY?: CSSProperties["overflowY"];
  b?: CSSProperties["border"];
  bw?: CSSProperties["borderWidth"];
  bs?: CSSProperties["borderStyle"];
  bc?: CSSProperties["borderColor"];
  br?: CSSProperties["borderRadius"];
  border?: CSSProperties["border"];
  borderWidth?: CSSProperties["borderWidth"];
  borderStyle?: CSSProperties["borderStyle"];
  borderColor?: CSSProperties["borderColor"];
  borderRadius?: CSSProperties["borderRadius"];
  inset?: SpacingValue;
  top?: SpacingValue;
  right?: SpacingValue;
  bottom?: SpacingValue;
  left?: SpacingValue;
  p?: SpacingValue;
  px?: SpacingValue;
  py?: SpacingValue;
  pt?: SpacingValue;
  pr?: SpacingValue;
  pb?: SpacingValue;
  pl?: SpacingValue;
  m?: SpacingValue;
  mx?: SpacingValue;
  my?: SpacingValue;
  mt?: SpacingValue;
  mr?: SpacingValue;
  mb?: SpacingValue;
  ml?: SpacingValue;
  flexBasis?: CSSProperties["flexBasis"];
  flexGrow?: CSSProperties["flexGrow"];
  flexShrink?: CSSProperties["flexShrink"];
  gridArea?: CSSProperties["gridArea"];
  gridColumn?: CSSProperties["gridColumn"];
  gridColumnStart?: CSSProperties["gridColumnStart"];
  gridColumnEnd?: CSSProperties["gridColumnEnd"];
  gridRow?: CSSProperties["gridRow"];
  gridRowStart?: CSSProperties["gridRowStart"];
  gridRowEnd?: CSSProperties["gridRowEnd"];
};